backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
openai==1.45.0
chromadb==0.5.23
pypdf==5.0.1
anthropic==0.30.0
flask-jwt-extended==4.6.0
bcrypt==4.2.1